MediaContact -> Related Media

+++ modules/RSNMediaContacts/models/Relation.php
@@ -5,12 +5,13 @@
 
 class RSNMediaContacts_Relation_Model extends Vtiger_Relation_Model{
 
-	// La suppression de la relation entre tables conduit ˆ la suppression de l'entit 
+	// La suppression de la relation entre tables conduit la suppression de l'entité
 	public function deleteRelation($sourceRecordId, $relatedRecordId){
 		if(parent::deleteRelation($sourceRecordId, $relatedRecordId)){
 			
 			$record = Vtiger_Record_Model::getInstanceById($relatedRecordId, 'RSNMediaRelations');
-			$record->delete();
+			if($record->getModule() == 'RSNMediaRelations')
+				$record->delete();
 			
 			return true;
 		}
